Move simulation-related code from headers to split header/source
