+++ source/kettle_internal/simulation/measure.cpp
@@ -1,28 +1,24 @@
-#pragma once
-
 #include <cstddef>
-#include <optional>
-#include <random>
 
-#include "kettle_internal/common/prng.hpp"
 #include "kettle_internal/common/utils_internal.hpp"
 #include "kettle/gates/primitive_gate.hpp"
 #include "kettle/state/state.hpp"
-#include "kettle/simulation/gate_pair_generator.hpp"
 
 #include "kettle_internal/gates/primitive_gate/gate_create.hpp"
+#include "kettle_internal/simulation/gate_pair_generator.hpp"
+#include "kettle_internal/simulation/measure.hpp"
 
-namespace impl_ket
+namespace ket::internal
 {
 
-inline auto probabilities_of_collapsed_states_(
+auto probabilities_of_collapsed_states_(
     ket::QuantumState& state,
     const ket::GateInfo& info
 ) -> std::tuple<double, double>
 {
     const auto target_index = ket::internal::create::unpack_single_qubit_gate_index(info);
 
-    auto pair_iterator = SingleQubitGatePairGenerator {target_index, state.n_qubits()};
+    auto pair_iterator = ket::internal::SingleQubitGatePairGenerator {target_index, state.n_qubits()};
     pair_iterator.set_state(0);
 
     auto prob_of_0_states = double {0.0};
@@ -47,7 +43,7 @@ void collapse_and_renormalize_(
 {
     const auto target_index = ket::internal::create::unpack_single_qubit_gate_index(info);
 
-    auto pair_iterator = SingleQubitGatePairGenerator {target_index, state.n_qubits()};
+    auto pair_iterator = ket::internal::SingleQubitGatePairGenerator {target_index, state.n_qubits()};
     pair_iterator.set_state(0);
 
     for (std::size_t i {0}; i < pair_iterator.size(); ++i) {
@@ -66,38 +62,17 @@ void collapse_and_renormalize_(
         }
     }
 }
-
-/*
-    Perform a measurement at the target qubit index, which collapses the state.
-
-    For the time being, this is only done with a single-threaded implementation, because the
-    threads for the multithreaded implementation are spawned before entering the simulation
-    loop.
-*/
-template <ket::internal::DiscreteDistribution Distribution = std::discrete_distribution<int>>
-auto simulate_measurement_(
+template
+void collapse_and_renormalize_<0>(
     ket::QuantumState& state,
     const ket::GateInfo& info,
-    std::optional<int> seed = std::nullopt
-) -> Distribution::result_type
-{
-    const auto [prob_of_0_states, prob_of_1_states] = probabilities_of_collapsed_states_(state, info);
+    double norm_of_surviving_state
+);
+template
+void collapse_and_renormalize_<1>(
+    ket::QuantumState& state,
+    const ket::GateInfo& info,
+    double norm_of_surviving_state
+);
 
-    auto prng = ket::internal::get_prng_(seed);
-    auto coin_flipper = Distribution {{prob_of_0_states, prob_of_1_states}};
-
-    const auto collapsed_state = coin_flipper(prng);
-
-    if (collapsed_state == 0) {
-        const auto norm = std::sqrt(1.0 / prob_of_0_states);
-        collapse_and_renormalize_<1>(state, info, norm);
-    }
-    else {
-        const auto norm = std::sqrt(1.0 / prob_of_1_states);
-        collapse_and_renormalize_<0>(state, info, norm);
-    }
-
-    return collapsed_state;
-}
-
-}  // namespace impl_ket
+}  // namespace ket::internal
