+++ source/kettle_internal/simulation/multithread_simulate_utils.cpp
@@ -1,18 +1,17 @@
-#pragma once
-
 #include <stdexcept>
 #include <vector>
 
-#include <kettle/simulation/simulate_utils.hpp>
+#include "kettle_internal/simulation/simulate_utils.hpp"
+#include "kettle_internal/simulation/multithread_simulate_utils.hpp"
 
 /*
     This header file contains code to help perform the multithread quantum circuit simulations.
 */
 
-namespace impl_ket
+namespace ket::internal
 {
 
-inline auto load_balanced_division_(std::size_t numerator, std::size_t denominator) -> std::vector<std::size_t>
+auto load_balanced_division_(std::size_t numerator, std::size_t denominator) -> std::vector<std::size_t>
 {
     if (denominator == 0) {
         throw std::runtime_error {"Cannot perform load balancing division among 0 threads"};
@@ -29,7 +28,7 @@ inline auto load_balanced_division_(std::size_t numerator, std::size_t denominat
     return output;
 }
 
-inline auto partial_sums_from_zero_(const std::vector<std::size_t>& values) -> std::vector<std::size_t>
+auto partial_sums_from_zero_(const std::vector<std::size_t>& values) -> std::vector<std::size_t>
 {
     auto output = std::vector<std::size_t> {};
     output.reserve(values.size() + 1);
@@ -42,7 +41,7 @@ inline auto partial_sums_from_zero_(const std::vector<std::size_t>& values) -> s
     return output;
 }
 
-inline auto partial_sum_pairs_(std::size_t n_gate_pairs, std::size_t n_threads) -> std::vector<FlatIndexPair>
+auto partial_sum_pairs_(std::size_t n_gate_pairs, std::size_t n_threads) -> std::vector<FlatIndexPair>
 {
     const auto gate_splits = load_balanced_division_(n_gate_pairs, n_threads);
 
@@ -59,4 +58,4 @@ inline auto partial_sum_pairs_(std::size_t n_gate_pairs, std::size_t n_threads) 
     return output;
 }
 
-}  // namespace impl_ket
+}  // namespace ket::internal
