+++ test/source/simulation/multithread_simulate_utils_test.cpp
@@ -6,7 +6,7 @@
 #include <catch2/matchers/catch_matchers_vector.hpp>
 #include <catch2/generators/catch_generators.hpp>
 
-#include <kettle/simulation/multithread_simulate_utils.hpp>
+#include "kettle_internal/simulation/multithread_simulate_utils.hpp"
 
 TEST_CASE("load_balanced_division()")
 {
@@ -26,14 +26,14 @@ TEST_CASE("load_balanced_division()")
             TestCase {32, 8, {4, 4, 4, 4, 4, 4, 4, 4}}
         );
 
-        const auto actual = impl_ket::load_balanced_division_(testcase.numerator, testcase.denominator);
+        const auto actual = ket::internal::load_balanced_division_(testcase.numerator, testcase.denominator);
 
         REQUIRE_THAT(actual, Catch::Matchers::UnorderedEquals(actual));
     }
 
     SECTION("throws when denominator is 0")
     {
-        REQUIRE_THROWS_AS(impl_ket::load_balanced_division_(10, 0), std::runtime_error);
+        REQUIRE_THROWS_AS(ket::internal::load_balanced_division_(10, 0), std::runtime_error);
     }
 }
 
@@ -53,7 +53,7 @@ TEST_CASE("partial_sums_from_zero()")
         TestCase { {2, 2, 2, 2}, {0, 2, 4, 6, 8} }
     );
 
-    const auto actual = impl_ket::partial_sums_from_zero_(testcase.values);
+    const auto actual = ket::internal::partial_sums_from_zero_(testcase.values);
 
     REQUIRE_THAT(actual, Catch::Matchers::Equals(testcase.expected));
 }
@@ -64,7 +64,7 @@ TEST_CASE("partial_sum_pairs_()")
     {
         std::size_t n_gate_pairs;
         std::size_t n_threads;
-        std::vector<impl_ket::FlatIndexPair> expected;
+        std::vector<ket::internal::FlatIndexPair> expected;
     };
 
     const auto testcase = GENERATE(
@@ -75,7 +75,7 @@ TEST_CASE("partial_sum_pairs_()")
         TestCase { 16, 4, {{0, 4}, {4, 8}, {8, 12}, {12, 16}} }
     );
 
-    const auto actual = impl_ket::partial_sum_pairs_(testcase.n_gate_pairs, testcase.n_threads);
+    const auto actual = ket::internal::partial_sum_pairs_(testcase.n_gate_pairs, testcase.n_threads);
 
     REQUIRE_THAT(actual, Catch::Matchers::Equals(testcase.expected));
 }
